refactor: ecosystem actor about
