+++ pages/ecosystem-actors/[code]/projects.tsx
@@ -4,7 +4,7 @@ import { fetchProjects } from '@ses/containers/ActorProjects/api/query';
 import { ResourceType } from '@ses/core/models/interfaces/types';
 import { featureFlags } from 'feature-flags/feature-flags';
 import React from 'react';
-import { fetchActorAbout } from '@/views/EAAbout/api/queries';
+import { fetchActorAbout } from '@/views/EcosystemActorAbout/api/queries';
 import { fetchActors } from '@/views/EcosystemActorsIndex/api/queries';
 import type { GetServerSideProps, GetServerSidePropsContext, InferGetServerSidePropsType, NextPage } from 'next';
