+++ src/stories/containers/TransparencyReport/components/TransparencyAuditorComments/ParticipantRoles.tsx
@@ -3,7 +3,7 @@ import { useThemeContext } from '@ses/core/context/ThemeContext';
 import { ResourceType } from '@ses/core/models/interfaces/types';
 import lightTheme from '@ses/styles/theme/themes';
 import React from 'react';
-import InlineUser from '../../../../../views/EAAbout/components/InlineUser/InlineUser';
+import InlineUser from '../../../../../views/EcosystemActorAbout/components/InlineUser/InlineUser';
 import type { BaseUser } from '@ses/core/models/interfaces/users';
 
 export type ParticipantRolesProps = {
