+++ src/views/EcosystemActorAbout/api/queries.ts
