+++ src/views/EcosystemActorAbout/components/ActorNavigationOptions/ActorNavigationOptions.tsx
