+++ src/views/EcosystemActorAbout/components/ActorSummary/ActorSummary.tsx
