+++ src/views/EcosystemActorAbout/components/ActorSummary/ActorTitleAbout.tsx
