+++ src/views/EcosystemActorAbout/components/ActorSummary/ActorTitleWithDescription.tsx
