+++ src/views/EcosystemActorAbout/components/InlineUser/InlineUser.tsx
