+++ src/views/EcosystemActorAbout/utils.ts
